pyproject.toml
[project]
name = "mcp_langgraph_tools"
version = "0.1.0"
description = "Mcp Langgraph Tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "langchain-anthropic>=0.1.23",
    "langchain-community>=0.2.12",
    "langchain-core>=0.2.34",
    "langchain-google-genai>=1.0.10",
    "langchain-groq>=0.1.9",
    "langchain-ollama>=0.1.1",
    "langchain-openai>=0.1.22",
    "langchain>=0.2.14",
    "langgraph>=0.2.14",
    "mcp>=1.0.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
]
packages = ["src/mcp_langgraph_tools"]

[project.scripts]
mcp_langgraph_tools = "mcp_langgraph_tools.__main__:main"

[tool.setuptools.package-data]
mcp_langgraph_tools = [
    "py.typed",
    "*/*.png",
    "*/*.md",
    "*/*.tcss",
    "*.png",
    "*.md",
    "*.tcss"
]

[build-system]
requires = ["hatchling", "wheel"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "build>=1.2.1",
    "twine>=5.1.1",
    "pyright>=1.1.379",
    "pre-commit>=3.8.0",
    "ruff>=0.7.0",
    "types-orjson>=3.6.2",
    "pyinstrument>=5.0.0",
]

[tool.hatch.version]
path = "src/mcp_langgraph_tools/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_langgraph_tools"]
include = [
    "*.py",
    "py.typed",
    "*.png",
    "*.md",
    "*.tcss",
    "*.png",
    "*.md",
    "*.tcss"
]

[tool.hatch.build.targets.sdist]
include = [
    "src/mcp_langgraph_tools",
    "LICENSE",
    "README.md",
    "pyproject.toml"
]
exclude = [
    "*.pyc",
    "__pycache__",
    "*.so",
    "*.dylib"
]
